feat: 优化Agent超时导致的任务无法正常启动 #11546

+++ src/backend/ci/core/process/biz-engine/src/main/kotlin/com/tencent/devops/process/engine/atom/vm/DispatchVMStartupTaskAtom.kt
@@ -399,7 +399,7 @@ class DispatchVMStartupTaskAtom @Autowired constructor(
                 try {
                     thirdPartyAgentRollBackMonitor(task)
                     // 修改轮训时间,3m30s方便命中dispath的3分钟频率
-                    task.taskParams[BS_ATOM_STATUS_REFRESH_DELAY_MILLS] = 3.5 * 60 * 1000
+                    task.taskParams[BS_ATOM_STATUS_REFRESH_DELAY_MILLS] = (3.5 * 60 * 1000).toInt()
                 } catch (ignore: Exception) {
                     // 忽略掉因调用打印接口出错而导致调度失败的问题
                 }
